separate out complex constraint readers to their own class
