+++ profile/src/main/java/com/scottlogic/deg/profile/reader/AtomicConstraintTypeReaderMap.java
@@ -19,13 +19,13 @@ package com.scottlogic.deg.profile.reader;
 import com.google.inject.Inject;
 import com.scottlogic.deg.common.ValidationException;
 import com.scottlogic.deg.common.profile.Field;
-import com.scottlogic.deg.common.profile.constraintdetail.ParsedDateGranularity;
-import com.scottlogic.deg.common.profile.constraintdetail.ParsedGranularity;
 import com.scottlogic.deg.common.profile.constraints.atomic.*;
 import com.scottlogic.deg.common.util.Defaults;
 import com.scottlogic.deg.generator.fieldspecs.whitelist.DistributedSet;
 import com.scottlogic.deg.generator.fieldspecs.whitelist.WeightedElement;
 import com.scottlogic.deg.generator.fieldspecs.whitelist.FrequencyDistributedSet;
+import com.scottlogic.deg.profile.reader.constraintreaders.FromFileReader;
+import com.scottlogic.deg.profile.reader.constraintreaders.GranularToReader;
 import com.scottlogic.deg.profile.reader.constraintreaders.OfTypeReader;
 import com.scottlogic.deg.profile.reader.file.CsvInputStreamReader;
 import com.scottlogic.deg.profile.v0_1.AtomicConstraintType;
@@ -35,7 +35,6 @@ import java.math.BigDecimal;
 import java.time.OffsetDateTime;
 import java.util.HashMap;
 import java.util.Map;
-import java.util.Optional;
 import java.util.regex.Pattern;
 import java.util.stream.Collectors;
 
@@ -47,7 +46,6 @@ public class AtomicConstraintTypeReaderMap {
 
     private final String fromFilePath;
 
-    @Inject
     public AtomicConstraintTypeReaderMap(final String fromFilePath) {
         this.fromFilePath = fromFilePath;
     }
@@ -58,6 +56,8 @@ public class AtomicConstraintTypeReaderMap {
         Map<AtomicConstraintType, ConstraintReader> map = new HashMap<>();
 
         map.put(IS_OF_TYPE, new OfTypeReader());
+        map.put(IS_GRANULAR_TO, new GranularToReader());
+        map.put(IS_FROM_FILE, new FromFileReader(fromFilePath));
 
         map.put(FORMATTED_AS,
             (dto, fields) -> new FormatConstraint(
@@ -135,40 +135,6 @@ public class AtomicConstraintTypeReaderMap {
                     fields.getByName(dto.field),
                     getValidatedValue(dto, OffsetDateTime.class)));
 
-        map.put(IS_GRANULAR_TO,
-            (dto, fields) ->
-            {
-                Optional<Number> numberValidatedValue =
-                    ConstraintReaderHelpers.tryGetValidatedValue(dto, Number.class);
-                Optional<String> stringValidatedValue =
-                    ConstraintReaderHelpers.tryGetValidatedValue(dto, String.class);
-
-                if (numberValidatedValue.isPresent()) {
-                    Optional<ParsedGranularity> parsedNumericGranularity =
-                        ParsedGranularity.tryParse(numberValidatedValue.get());
-                    if (parsedNumericGranularity.isPresent()) {
-                        return new IsGranularToNumericConstraint(
-                            fields.getByName(dto.field),
-                            parsedNumericGranularity.get()
-                        );
-                    }
-                } else if (stringValidatedValue.isPresent()) {
-                    Optional<ParsedDateGranularity> parsedDateGranularity =
-                        ParsedDateGranularity.tryParse(stringValidatedValue.get());
-                    if (parsedDateGranularity.isPresent()) {
-                        return new IsGranularToDateConstraint(
-                            fields.getByName(dto.field),
-                            parsedDateGranularity.get()
-                        );
-                    }
-                }
-                throw new InvalidProfileException(String.format(
-                    "Field [%s]: Couldn't recognise granularity value, it must be either a negative power of ten or one of the supported datetime units.",
-                    dto.field
-                ));
-            }
-        );
-
         map.put(IS_NULL,
             (dto, fields) -> new IsNullConstraint(fields.getByName(dto.field)));
 
@@ -202,44 +168,8 @@ public class AtomicConstraintTypeReaderMap {
                         BigDecimal.ZERO,
                         maxStringLength)));
 
-        map.put(IS_FROM_FILE,
-            (dto, fields) -> {
-                String value = getValidatedValue(dto, String.class);
-
-                InputStream streamFromPath = createStreamFromPath(appendPath(value));
-                DistributedSet<String> names = CsvInputStreamReader.retrieveLines(streamFromPath);
-                closeStream(streamFromPath);
-
-                DistributedSet<Object> downcastedNames = new FrequencyDistributedSet<>(
-                    names.distributedSet().stream()
-                        .map(holder -> new WeightedElement<>((Object) holder.element(), holder.weight()))
-                        .collect(Collectors.toSet()));
-                Field field = fields.getByName(dto.field);
-
-                return new IsInSetConstraint(field, downcastedNames);
-            });
-
             return map;
     }
 
-    private String appendPath(String path) {
-        return fromFilePath != null ? fromFilePath + path : path;
-    }
-
-    private static InputStream createStreamFromPath(String path) {
-        try {
-            return new FileInputStream(path);
-        } catch (FileNotFoundException e) {
-            throw new ValidationException(e.getMessage());
-        }
-    }
-
-    private static void closeStream(InputStream stream) {
-        try {
-            stream.close();
-        } catch (IOException e) {
-            throw new UncheckedIOException(e);
-        }
-    }
 
 }
